+++ Runtime/Scripts/TextureLoaderBase.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) 2019-2021 Andreas Atteneder, All Rights Reserved.
+﻿// Copyright (c) 2019-2022 Andreas Atteneder, All Rights Reserved.
 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -22,11 +22,7 @@ namespace KtxUnity {
         /// </summary>
         /// <param name="result">Result of loading a texture</param>
         protected void OnTextureLoaded(TextureResult result) {
-            if(result!=null) {
-                ApplyTexture(result);
-            } else {
-                Debug.LogError("Loading Texture failed!");
-            }
+            ApplyTexture(result);
         }
 
         protected abstract void ApplyTexture(TextureResult result);