+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Error Codes (in `TextureResult.errorCode`)
 ### Changed
 - In release builds there's no console logging anymore (use the `errorCode` instead). In Debug builds and the Editor you still get detailed error messages.
+### Fixed
+- Will not transcode textures with sizes that are not a multiple of four to incompatible DXT5 or BC7 formats anymore
 
 ## [1.1.2] - 2021-10-27
 ### Added