+++ CHANGELOG.md
@@ -13,6 +13,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - WebGL library is built with Emscripten 2.0.19 now
 - Minimum required version is Unity 2021.2
 
+## [1.2.0] - 2022-04-14
+### Added
+- Editor Import via `ScriptableImporter`
+- Error Codes (in `TextureResult.errorCode`)
+### Changed
+- In release builds there's no console logging anymore (use the `errorCode` instead). In Debug builds and the Editor you still get detailed error messages.
+### Fixed
+- Will not transcode textures with sizes that are not a multiple of four to incompatible DXT5 or BC7 formats anymore
+- Re-compiled macOS native library in release mode (`MinSizeRel`; was `Debug`). Expect improved performance.
+
 ## [1.1.2] - 2021-10-27
 ### Added
 - Error message when users try to run KtxUnity 1.x Unity >=2021.2 combination targeting WebGL